Video-related fixes
